Agent: secure FS IPC, refactor editor & diff viewer
Add security and stability checks to agent IPC (assertWithinCodeDirectory to confine file ops to registered project code dirs; validate cwd on all platforms; ensure PTYs are killed when renderer is destroyed). Switch preload agent API to use the common invoke wrapper with typed returns and keep picker calls as raw invokes. Refactor editor/diff UI: extract FileEditorInner, ImageViewer and editorTheme, add a new DiffFile component (split/unified diff rendering, palettes, syntax highlighting) and update DiffViewer to use it. Improve terminal mount (guard against unmount during dynamic imports and honor font-scale) and add divider refs in AgentView for more robust resizing. Overall: security hardening, componentization, and reliability fixes.
